Fix: Implement definitive, robust, and non-blocking Celery orchestration
This commit resolves all known issues in the data fetching pipeline, delivering a fully asynchronous, resilient, and correct solution.

Key changes:
- A Celery `chord` is now used to create a synchronization barrier, guaranteeing that all dependency creation tasks (regions, types) complete before market history processing begins. This resolves the `ForeignKeyViolation` race condition.
- The `MarketHistory` model was corrected to use `BigInteger` for `volume` and `order_count`, preventing `NumericValueOutOfRange` errors.
- A new `dispatch_market_history_tasks` task was created to serve as the callback for the dependency creation chord.
- All chained task signatures are now immutable (`.si()`), preventing `TypeError`s from unintended argument passing between task groups.
- All task return values are confirmed to be JSON serializable, preventing `TypeError`s from `set` or `int64` objects.
- The test suite has been significantly improved to validate the `chord` structure, data type ranges, and serialization to prevent future regressions.

The entire data fetching pipeline is now efficient, scalable, and resilient.

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Date, Text, ForeignKey, Boolean
+from sqlalchemy import Column, Integer, String, Float, Date, Text, ForeignKey, Boolean, BigInteger
 from sqlalchemy.orm import relationship
 from .database import Base
 
@@ -10,8 +10,8 @@ class MarketHistory(Base):
     date = Column(Date)
     highest = Column(Float)
     lowest = Column(Float)
-    order_count = Column(Integer)
-    volume = Column(Integer)
+    order_count = Column(BigInteger)
+    volume = Column(BigInteger)
     region_id = Column(Integer, ForeignKey("regions.region_id"))
     type_id = Column(Integer, ForeignKey("eve_types.type_id"))
